Update README.md
In the privacy policy we state how Ojala collects, uses, stores, and discloses users' personal information when users access the Ojala service. Personal information can include anything from names and email addresses to form submissions.
retrieval.

The privacy policy provides users with information about their rights regarding their personal data when using Ojala. 
This is a comprehensive privacy policy that clearly communicates privacy practices for Ojala, so that users are not only aware of the how all their data is used but also how it is best used to serve their purposes while using Ojala

The Privacy Policy page for Ojala is a critical document that ensures our users are fully informed about how their personal information is handled so that users not only feel comfortable using the Ojala platform but also to enhance the user experience of the users that visit and use the Ojala platform and/or service.